+++ dcli/lib/src/functions/sleep.dart
@@ -12,6 +12,10 @@ import '../../dcli.dart';
 
 /// sleeps for the provided [duration] of the given [interval].
 ///
+/// WARINING: call this function will stop all async code from
+/// running. This is a change from 3.x and previous version of dcli.
+/// Use [sleepAsync] if you need async code to keep running.
+///
 /// ```dart
 /// sleep(2);
 ///
@@ -21,6 +25,9 @@ import '../../dcli.dart';
 /// The [interval] defaults to seconds.
 ///
 /// If the duration is 0 or less sleep returns immediately.
+///
+/// See: [sleepAsync]
+///
 void sleep(int duration, {Interval interval = Interval.seconds}) =>
     _Sleep().sleep(duration, interval: interval);
 
@@ -39,6 +46,24 @@ enum Interval {
   /// the duration argument is in seconds
   minutes
 }
+
+/// sleeps for the provided [duration] of the given [interval].
+///
+/// ```dart
+/// sleepAsync(2);
+///
+/// sleepAsync(2, interval=Interval.minutes);
+/// ```
+///
+/// The [interval] defaults to seconds.
+///
+/// If the duration is 0 or less sleep returns immediately.
+///
+/// See: [sleep]
+///
+Future<void> sleepAsync(int duration,
+        {Interval interval = Interval.seconds}) async =>
+    _Sleep().sleepAsync(duration, interval: interval);
 
 class _Sleep extends core.DCliFunction {
   void sleep(int duration, {Interval interval = Interval.seconds}) {
@@ -60,5 +85,27 @@ class _Sleep extends core.DCliFunction {
     }
 
     io.sleep(duration0);
+  }
+
+  Future<void> sleepAsync(int duration,
+      {Interval interval = Interval.seconds}) async {
+    verbose(() => 'sleep: duration: $duration interval: $interval');
+    late Duration duration0;
+    switch (interval) {
+      case Interval.hours:
+        duration0 = Duration(hours: duration);
+        break;
+      case Interval.seconds:
+        duration0 = Duration(seconds: duration);
+        break;
+      case Interval.milliseconds:
+        duration0 = Duration(milliseconds: duration);
+        break;
+      case Interval.minutes:
+        duration0 = Duration(minutes: duration);
+        break;
+    }
+
+    Future.delayed(duration0, () {});
   }
 }